+++ core/src/main/java/io/devarium/core/auth/OAuth2Client.java
@@ -1,0 +1,6 @@
+package io.devarium.core.auth;
+
+public interface OAuth2Client {
+
+    OAuth2UserInfo getUserInfo(String code);
+}
