+++ core/src/main/java/io/devarium/core/auth/OAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package io.devarium.core.domain.user;
+package io.devarium.core.auth;
 
 public record OAuth2UserInfo(
     String id,
