+++ infrastructure/src/main/java/io/devarium/infrastructure/auth/oauth2/GoogleOAuth2Client.java
@@ -2,11 +2,11 @@ package io.devarium.infrastructure.auth.oauth2;
 
 import static io.devarium.infrastructure.auth.jwt.JwtConstants.BEARER_PREFIX;
 
+import io.devarium.core.auth.OAuth2Client;
+import io.devarium.core.auth.OAuth2Provider;
+import io.devarium.core.auth.OAuth2UserInfo;
 import io.devarium.core.auth.exception.AuthErrorCode;
 import io.devarium.core.auth.exception.CustomAuthException;
-import io.devarium.core.domain.user.OAuth2Provider;
-import io.devarium.core.domain.user.OAuth2UserInfo;
-import io.devarium.core.domain.user.port.OAuth2Client;
 import io.devarium.infrastructure.auth.oauth2.dto.GoogleTokenResponse;
 import io.devarium.infrastructure.auth.oauth2.dto.GoogleUserInfoResponse;
 import lombok.RequiredArgsConstructor;
